making arviz an optional dep
