+++ .github/workflows/tests.yml
@@ -13,12 +13,20 @@ jobs:
     runs-on: "ubuntu-latest"
     strategy:
       matrix:
-        python-version: [3.9]
+        python-version: ["3.9"]
         arviz-version:
+          - ""
           - "arviz~=0.9"
           - "arviz~=0.10"
           - "arviz~=0.11"
           - "https://github.com/arviz-devs/arviz/archive/main.zip"
+        include:
+          - python-version: "3.6"
+            arviz-version: "arviz"
+          - python-version: "3.7"
+            arviz-version: "arviz"
+          - python-version: "3.8"
+            arviz-version: "arviz"
 
     steps:
       - uses: actions/checkout@v2
@@ -31,8 +39,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install -U pip
-          python -m pip install $ARVIZ
-          python -m pip install ".[test]"
+          python -m pip install $ARVIZ ".[test]"
         env:
           ARVIZ: ${{ matrix.arviz-version }}
